pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biologger-sim"
version = "0.1.0"
description = "Real-time simulation environment for marine animal tracking and behavioral prediction"
authors = [
    {name = "Daniel Fry"},
    {name = "Camrin Braun"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "pyzmq",
    "pydantic",
    "PyYAML"
]

[project.scripts]
biologger-sim = "biologger_sim.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["biologger_sim*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "--color=yes"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["biologger_sim"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
